add a substructuring mechanism
leverage section-tags to create substructur in your presentation.

ATM this is still a bit awkward, you have to create a structure like
the following:

  <section class="structure id="Motivation">
    <section class="slides">
      <article>
	...
      </article>
    </section>
  <section>
  <section class="structure id="Approach">
    <section class="slides">
    ...
    </section>
  </section>
  <section class="structure" id="Evaluation>
  ...
  </section>
  <section class="structure" id="Conclusion>
  ...
  </section>

Note: The repetition of the inner section of class `slides` is
redundant and is an implementation artifact.

Note: You don't have to specify an outer section of class structure.

You can also create a table of contents like this:

  <section class="slides">
    <article>
      <ul class="theToc">
      </ul>
    </article>
  </section>